+++ smartcontract/cli/src/multicastgroup/create.rs
@@ -1,11 +1,12 @@
 use crate::{
     doublezerocommand::CliCommand,
-    poll_for_activation::poll_for_multicastgroup_activated,
     requirements::{CHECK_BALANCE, CHECK_ID_JSON},
     validators::{validate_code, validate_parse_bandwidth, validate_pubkey},
 };
 use clap::Args;
-use doublezero_sdk::commands::multicastgroup::create::CreateMulticastGroupCommand;
+use doublezero_sdk::commands::multicastgroup::{
+    create::CreateMulticastGroupCommand, get::GetMulticastGroupCommand,
+};
 use solana_sdk::pubkey::Pubkey;
 use std::{io::Write, str::FromStr};
 
@@ -46,8 +47,10 @@ impl CreateMulticastGroupCliCommand {
         writeln!(out, "Signature: {signature}",)?;
 
         if self.wait {
-            let user = poll_for_multicastgroup_activated(client, &pubkey)?;
-            writeln!(out, "Status: {0}", user.status)?;
+            let (_, mgroup) = client.get_multicastgroup(GetMulticastGroupCommand {
+                pubkey_or_code: pubkey.to_string(),
+            })?;
+            writeln!(out, "Status: {0}", mgroup.status)?;
         }
 
         Ok(())