+++ smartcontract/cli/src/user/create_subscribe.rs
@@ -1,15 +1,15 @@
 use crate::{
     doublezerocommand::CliCommand,
     helpers::parse_pubkey,
-    poll_for_activation::poll_for_user_activated,
     requirements::{CHECK_BALANCE, CHECK_ID_JSON},
     validators::validate_pubkey_or_code,
 };
 use clap::Args;
 use doublezero_sdk::{
     commands::{
-        device::get::GetDeviceCommand, multicastgroup::get::GetMulticastGroupCommand,
-        user::create_subscribe::CreateSubscribeUserCommand,
+        device::get::GetDeviceCommand,
+        multicastgroup::get::GetMulticastGroupCommand,
+        user::{create_subscribe::CreateSubscribeUserCommand, get::GetUserCommand},
     },
     *,
 };
@@ -109,7 +109,7 @@ impl CreateSubscribeUserCliCommand {
         writeln!(out, "Signature: {signature}",)?;
 
         if self.wait {
-            let user = poll_for_user_activated(client, &pubkey)?;
+            let (_, user) = client.get_user(GetUserCommand { pubkey })?;
             writeln!(out, "Status: {0}", user.status)?;
         }